splitting of the segReadWrite unit tests
