+++ .gitignore
@@ -227,3 +227,5 @@ install_manifest.txt
 CPackConfig.cmake
 CPackSourceConfig.cmake
 *_CPack_Packages/
+/unit-tests/SegmentReadWriteLT
+/unit-tests/SegmentReadWriteT
