+++ unit-tests/SegmentReadWriteT.cc
@@ -158,353 +158,11 @@ public:
 
     } // end seg_write_test_0()
 
-    void seg_read_test_1() {
-        // setup test file
-        uint64_t bytes = std::pow(2, 31);
-        string filepath = "/tmp/srwt/segReadWriteTest1.bin";
-        int code = gen_large_test_file(bytes, filepath);
-        if (code != 0) {
-            CPPUNIT_FAIL("dd command failed with return code: " + to_string(code));
-        }
-
-        // call seg_read on file
-        vector<char> buff(bytes);
-        std::filebuf fb;
-        if (fb.open(filepath, std::ios::in)) {
-            std::istream is(&fb);
-            segmented_read(is, buff.data(), bytes);
-            fb.close();
-        }
-
-        // take seg_read data and write to file
-        string filepath2 = "/tmp/srwt/segReadWriteTest2.bin";
-        write_buffer_to_file(buff.data(), bytes, filepath2);
-
-        // assert (compare files)
-        auto command = "cmp " + filepath + " " + filepath2;
-        DBG(cerr << "Running command: " << command << endl);
-        int return_code = std::system(command.c_str());
-        CPPUNIT_ASSERT(return_code == 0);
-    } // end seg_read_test_1()
-
-    void seg_read_test_2() {
-        // setup test file
-        uint64_t bytes = std::pow(2, 31) - 1;
-        string filepath = "/tmp/srwt/segReadWriteTest1.bin";
-        int code = gen_large_test_file(bytes, filepath);
-        if (code != 0) {
-            CPPUNIT_FAIL("dd command failed with return code: " + to_string(code));
-        }
-
-        // call seg_read on file
-        vector<char> buff(bytes);
-        std::filebuf fb;
-        if (fb.open(filepath, std::ios::in)) {
-            std::istream is(&fb);
-            segmented_read(is, buff.data(), bytes);
-            fb.close();
-        }
-
-        // take seg_read data and write to file
-        string filepath2 = "/tmp/srwt/segReadWriteTest2.bin";
-        write_buffer_to_file(buff.data(), bytes, filepath2);
-
-        // assert (compare files)
-        auto command = "cmp " + filepath + " " + filepath2;
-        DBG(cerr << "Running command: " << command << endl);
-        int return_code = std::system(command.c_str());
-        CPPUNIT_ASSERT(return_code == 0);
-    } // end seg_read_test_2()
-
-    void seg_read_test_3() {
-        // setup test file
-        uint64_t bytes = std::pow(2, 31) + 1;
-        string filepath = "/tmp/srwt/segReadWriteTest1.bin";
-        int code = gen_large_test_file(bytes, filepath);
-        if (code != 0) {
-            CPPUNIT_FAIL("dd command failed with return code: " + to_string(code));
-        }
-
-        // call seg_read on file
-        vector<char> buff(bytes);
-        std::filebuf fb;
-        if (fb.open(filepath, std::ios::in)) {
-            std::istream is(&fb);
-            segmented_read(is, buff.data(), bytes);
-            fb.close();
-        }
-
-        // take seg_read data and write to file
-        string filepath2 = "/tmp/srwt/segReadWriteTest2.bin";
-        write_buffer_to_file(buff.data(), bytes, filepath2);
-
-        // assert (compare files)
-        auto command = "cmp " + filepath + " " + filepath2;
-        DBG(cerr << "Running command: " << command << endl);
-        int return_code = std::system(command.c_str());
-        CPPUNIT_ASSERT(return_code == 0);
-    } // end seg_read_test_3()
-
-    void seg_write_test_1() {
-        uint64_t bytes = std::pow(2, 31);
-        string filepath = "/tmp/srwt/segReadWriteTest1.bin";
-        int code = gen_large_test_file(bytes, filepath);
-        if (code != 0) {
-            CPPUNIT_FAIL("dd command failed with return code: " + to_string(code));
-        }
-
-        string filepath2 = "/tmp/srwt/segReadWriteTest2.bin";
-        vector<char> buffer = read_file_to_buffer(filepath);
-        filebuf fb;
-        if (fb.open(filepath2, ios::out)) {
-            ostream os(&fb);
-            segmented_write(os, buffer.data(), bytes);
-            fb.close();
-        }
-
-        // assert (compare files)
-        auto command = "cmp " + filepath + " " + filepath2;
-        DBG(cerr << "Running command: " << command << endl);
-        int return_code = system(command.c_str());
-        CPPUNIT_ASSERT(return_code == 0);
-
-    } // end seg_write_test_1()
-
-    void seg_write_test_2() {
-        uint64_t bytes = std::pow(2, 31) - 1;
-        string filepath = "/tmp/srwt/segReadWriteTest1.bin";
-        int code = gen_large_test_file(bytes, filepath);
-        if (code != 0) {
-            CPPUNIT_FAIL("dd command failed with return code: " + to_string(code));
-        }
-
-        string filepath2 = "/tmp/srwt/segReadWriteTest2.bin";
-        vector<char> buffer = read_file_to_buffer(filepath);
-        filebuf fb;
-        if (fb.open(filepath2, ios::out)) {
-            ostream os(&fb);
-            segmented_write(os, buffer.data(), bytes);
-            fb.close();
-        }
-
-        // assert (compare files)
-        auto command = "cmp " + filepath + " " + filepath2;
-        DBG(cerr << "Running command: " << command << endl);
-        int return_code = system(command.c_str());
-        CPPUNIT_ASSERT(return_code == 0);
-
-    } // end seg_write_test_2()
-
-    void seg_write_test_3() {
-        uint64_t bytes = std::pow(2, 31) + 1;
-        string filepath = "/tmp/srwt/segReadWriteTest1.bin";
-        int code = gen_large_test_file(bytes, filepath);
-        if (code != 0) {
-            CPPUNIT_FAIL("dd command failed with return code: " + to_string(code));
-        }
-
-        string filepath2 = "/tmp/srwt/segReadWriteTest2.bin";
-        vector<char> buffer = read_file_to_buffer(filepath);
-        filebuf fb;
-        if (fb.open(filepath2, ios::out)) {
-            ostream os(&fb);
-            segmented_write(os, buffer.data(), bytes);
-            fb.close();
-        }
-
-        // assert (compare files)
-        auto command = "cmp " + filepath + " " + filepath2;
-        DBG(cerr << "Running command: " << command << endl);
-        int return_code = system(command.c_str());
-        CPPUNIT_ASSERT(return_code == 0);
-
-    } // end seg_write_test_3()
-
-    void seg_read_test_4() {
-        // setup test file
-        uint64_t bytes = std::pow(2, 32);
-        string filepath = "/tmp/srwt/segReadWriteTest1.bin";
-        int code = gen_large_test_file(bytes, filepath);
-        if (code != 0) {
-            CPPUNIT_FAIL("dd command failed with return code: " + to_string(code));
-        }
-
-        // call seg_read on file
-        vector<char> buff(bytes);
-        std::filebuf fb;
-        if (fb.open(filepath, std::ios::in)) {
-            std::istream is(&fb);
-            segmented_read(is, buff.data(), bytes);
-            fb.close();
-        }
-
-        // take seg_read data and write to file
-        string filepath2 = "/tmp/srwt/segReadWriteTest2.bin";
-        write_buffer_to_file(buff.data(), bytes, filepath2);
-
-        // assert (compare files)
-        auto command = "cmp " + filepath + " " + filepath2;
-        DBG(cerr << "Running command: " << command << endl);
-        int return_code = std::system(command.c_str());
-        CPPUNIT_ASSERT(return_code == 0);
-    } // end seg_read_test_4()
-
-    void seg_read_test_5() {
-        // setup test file
-        uint64_t bytes = std::pow(2, 32) - 1;
-        string filepath = "/tmp/srwt/segReadWriteTest1.bin";
-        int code = gen_large_test_file(bytes, filepath);
-        if (code != 0) {
-            CPPUNIT_FAIL("dd command failed with return code: " + to_string(code));
-        }
-
-        // call seg_read on file
-        vector<char> buff(bytes);
-        std::filebuf fb;
-        if (fb.open(filepath, std::ios::in)) {
-            std::istream is(&fb);
-            segmented_read(is, buff.data(), bytes);
-            fb.close();
-        }
-
-        // take seg_read data and write to file
-        string filepath2 = "/tmp/srwt/segReadWriteTest2.bin";
-        write_buffer_to_file(buff.data(), bytes, filepath2);
-
-        // assert (compare files)
-        auto command = "cmp " + filepath + " " + filepath2;
-        DBG(cerr << "Running command: " << command << endl);
-        int return_code = std::system(command.c_str());
-        CPPUNIT_ASSERT(return_code == 0);
-    } // end seg_read_test_5()
-
-    void seg_read_test_6() {
-        // setup test file
-        uint64_t bytes = std::pow(2, 32) + 1;
-        string filepath = "/tmp/srwt/segReadWriteTest1.bin";
-        int code = gen_large_test_file(bytes, filepath);
-        if (code != 0) {
-            CPPUNIT_FAIL("dd command failed with return code: " + to_string(code));
-        }
-
-        // call seg_read on file
-        vector<char> buff(bytes);
-        std::filebuf fb;
-        if (fb.open(filepath, std::ios::in)) {
-            std::istream is(&fb);
-            segmented_read(is, buff.data(), bytes);
-            fb.close();
-        }
-
-        // take seg_read data and write to file
-        string filepath2 = "/tmp/srwt/segReadWriteTest2.bin";
-        write_buffer_to_file(buff.data(), bytes, filepath2);
-
-        // assert (compare files)
-        auto command = "cmp " + filepath + " " + filepath2;
-        DBG(cerr << "Running command: " << command << endl);
-        int return_code = std::system(command.c_str());
-        CPPUNIT_ASSERT(return_code == 0);
-    } // end seg_read_test_6()
-
-    void seg_write_test_4() {
-        uint64_t bytes = std::pow(2, 32);
-        string filepath = "/tmp/srwt/segReadWriteTest1.bin";
-        int code = gen_large_test_file(bytes, filepath);
-        if (code != 0) {
-            CPPUNIT_FAIL("dd command failed with return code: " + to_string(code));
-        }
-
-        string filepath2 = "/tmp/srwt/segReadWriteTest2.bin";
-        vector<char> buffer = read_file_to_buffer(filepath);
-        filebuf fb;
-        if (fb.open(filepath2, ios::out)) {
-            ostream os(&fb);
-            segmented_write(os, buffer.data(), bytes);
-            fb.close();
-        }
-
-        // assert (compare files)
-        auto command = "cmp " + filepath + " " + filepath2;
-        DBG(cerr << "Running command: " << command << endl);
-        int return_code = system(command.c_str());
-        CPPUNIT_ASSERT(return_code == 0);
-
-    } // end seg_write_test_4()
-
-    void seg_write_test_5() {
-        uint64_t bytes = std::pow(2, 32) - 1;
-        string filepath = "/tmp/srwt/segReadWriteTest1.bin";
-        int code = gen_large_test_file(bytes, filepath);
-        if (code != 0) {
-            CPPUNIT_FAIL("dd command failed with return code: " + to_string(code));
-        }
-
-        string filepath2 = "/tmp/srwt/segReadWriteTest2.bin";
-        vector<char> buffer = read_file_to_buffer(filepath);
-        filebuf fb;
-        if (fb.open(filepath2, ios::out)) {
-            ostream os(&fb);
-            segmented_write(os, buffer.data(), bytes);
-            fb.close();
-        }
-
-        // assert (compare files)
-        auto command = "cmp " + filepath + " " + filepath2;
-        DBG(cerr << "Running command: " << command << endl);
-        int return_code = system(command.c_str());
-        CPPUNIT_ASSERT(return_code == 0);
-
-    } // end seg_write_test_5()
-
-    void seg_write_test_6() {
-        uint64_t bytes = std::pow(2, 32) + 1;
-        string filepath = "/tmp/srwt/segReadWriteTest1.bin";
-        int code = gen_large_test_file(bytes, filepath);
-        if (code != 0) {
-            CPPUNIT_FAIL("dd command failed with return code: " + to_string(code));
-        }
-
-        string filepath2 = "/tmp/srwt/segReadWriteTest2.bin";
-        vector<char> buffer = read_file_to_buffer(filepath);
-        filebuf fb;
-        if (fb.open(filepath2, ios::out)) {
-            ostream os(&fb);
-            segmented_write(os, buffer.data(), bytes);
-            fb.close();
-        }
-
-        // assert (compare files)
-        auto command = "cmp " + filepath + " " + filepath2;
-        DBG(cerr << "Running command: " << command << endl);
-        int return_code = system(command.c_str());
-        CPPUNIT_ASSERT(return_code == 0);
-
-    } // end seg_write_test_6()
-
     CPPUNIT_TEST_SUITE(SegmentReadWriteT);
 
-    // Baby tests
+    // tests
     CPPUNIT_TEST(seg_read_test_0);
     CPPUNIT_TEST(seg_write_test_0);
-
-    // 2GB tests
-    CPPUNIT_TEST(seg_read_test_1); // seg_read test 2^31
-    CPPUNIT_TEST(seg_read_test_2); // seg_read test 2^31 - 1
-    CPPUNIT_TEST(seg_read_test_3); // seg_read test 2^31 + 1
-
-    CPPUNIT_TEST(seg_write_test_1); // seg_write test 2^31
-    CPPUNIT_TEST(seg_write_test_2); // seg_write test 2^31 - 1
-    CPPUNIT_TEST(seg_write_test_3); // seg_write test 2^31 + 1
-
-    // 4GB tests
-    CPPUNIT_TEST(seg_read_test_4); // seg_read test 2^32
-    CPPUNIT_TEST(seg_read_test_5); // seg_read test 2^32 - 1
-    CPPUNIT_TEST(seg_read_test_6); // seg_read test 2^32 + 1
-
-    CPPUNIT_TEST(seg_write_test_4); // seg_write test 2^32
-    CPPUNIT_TEST(seg_write_test_5); // seg_write test 2^32 - 1
-    CPPUNIT_TEST(seg_write_test_6); // seg_write test 2^32 + 1
 
     CPPUNIT_TEST_SUITE_END();
 };
